refactor: clean up, move and remove unnecessary code
