+++ src/app/(drawer)/(settings)/_layout.tsx
@@ -1,6 +1,22 @@
-import { Stack } from "expo-router"
+import { router, Stack } from "expo-router"
+import { useEffect } from "react"
+import { BackHandler } from "react-native"
 
 export default function Layout() {
+	useEffect(() => {
+		const backAction = () => {
+			router.navigate("/(drawer)/(tabs)/home")
+			return true
+		}
+
+		const backHandler = BackHandler.addEventListener(
+			"hardwareBackPress",
+			backAction,
+		)
+
+		return () => backHandler.remove()
+	}, [])
+
 	return (
 		<Stack screenOptions={{ headerShown: false }}>
 			<Stack.Screen
