+++ src/app/(drawer)/(settings)/language.tsx
@@ -1,23 +1,8 @@
-import { Text, BackHandler } from "react-native"
-import React, { useEffect } from "react"
 import ScreenWrapper from "@/components/ScreenWrapper"
-import { router } from "expo-router"
+import React from "react"
+import { Text } from "react-native"
 
 const language = () => {
-	useEffect(() => {
-		const backAction = () => {
-			router.navigate("/(drawer)/(tabs)/home")
-			return true
-		}
-
-		const backHandler = BackHandler.addEventListener(
-			"hardwareBackPress",
-			backAction,
-		)
-
-		return () => backHandler.remove()
-	}, [])
-
 	return (
 		<ScreenWrapper>
 			<Text>language</Text>
