+++ src/app/(drawer)/(settings)/notifications.tsx
@@ -1,23 +1,8 @@
-import { View, Text, BackHandler } from "react-native"
-import React, { useEffect } from "react"
 import ScreenWrapper from "@/components/ScreenWrapper"
-import { router } from "expo-router"
+import React from "react"
+import { Text } from "react-native"
 
 const notifications = () => {
-	useEffect(() => {
-		const backAction = () => {
-			router.navigate("/(drawer)/(tabs)/home")
-			return true
-		}
-
-		const backHandler = BackHandler.addEventListener(
-			"hardwareBackPress",
-			backAction,
-		)
-
-		return () => backHandler.remove()
-	}, [])
-
 	return (
 		<ScreenWrapper>
 			<Text>notifications</Text>
